requirements.txt
selenium==4.23.1
google-generativeai==0.6.0
fake-useragent==1.5.1
tqdm==4.66.4
webdriver-manager==4.0.2
pyfiglet==1.0.2
scikit-learn==1.3.2
python-dotenv==1.0.0